Add game state, event handling, and constants modules
Introduce a modular structure for the game by creating separate files for constants, game state, event handling, and types. This refactor enhances code organization and maintainability. The constants file defines key game parameters, while the game state file manages the game's logic and state transitions. The event handler file processes user inputs and updates the game accordingly. This change lays the groundwork for further development and feature enhancements.
